Add a pre-commit hook that can be used to avoid committing binary shaders
This is a client side hook, so if you want it you need to add it yourself
(refer to the comment). It is useful to make sure you have stripped the NULL
byte from the end of shaders dumped with Helix, since that causes git to treat
them as binary files and won't display their contents for git diff, log, etc.