Betrayer: Add lake reflection pixel shader dumped from game
This one is broken - v5 is passed in from the 1C208DE9 vertex shader (where it
was o7) and is projected here. I can adjust the X coord easily enough, but it
isn't enough and if I multiply it to adjust more it becomes apparent that it
isn't working properly and the depth is inverted. I hazard a guess that it may
be because I need to adjust it based on the apparent depth of the reflection?

Reflections just worked in the Ocean shader, but they already had some stereo
corrections there - maybe I could compare it?