Betrayer: Fix reflections on lake
Finally figured it out! The trick was that I had to use r2.z as the depth
instead of r2.w as I had first assumed. I'm guessing this was to do with
needing the depth of the object being reflected as opposed to the depth of the
water, but I'm not sure.

Whatever the case, the reflections are now 100% perfect and this lake looks
absolutely stunning in 3D =D