Betrayer: Remove crosshair vertex shader override
I assume this was the crosshair vertex shader anyway - it's changed in the
game, so this doesn't work on current versions anyway.

From the notes I took this is probably now vertex shader 1AFAC636 if anyone
wants to re-implement the RightMouseScript stuff to toggle the crosshair on and
off automatically.

This also removes the corresponding settings from the dx9settings.ini - my fix
pushes the entire UI back, so I don't need to match the texture and I'm not
using RightMouseScript so I don't need to handle the keys it sends (which
conflict with the default keybindings of the game anyway). If someone prefers
the way Eqzitara did it I'm happy to accept a pull request re-adding that to
work with the latest version of the game, but for now it's easier to merge if I
just remove it.