Betrayer: Add two more shaders that use TextureSpaceBlurOrigin
Because the lightshaft fix isn't perfect yet - the shafts are good, but the sun
in the otherworld is smeared and the shafts extend to uncomfortably different
lengths in each eye. Since these shaders use the same input as the light
shafts, it stands to reason that they are broken in exactly the same way.