Betrayer: Completely fix light shafts
Hooray! This game might actually be glitch free now!

Removed the ability to disable light shafts - why would you want to disable
that which is perfect?

Need to check if there is still any issues with the enemies, but they are very
uncooperative and move around a lot, so it's hard to tell.

I'll have to play through once more to be sure - there was one minor broken
effect near the end (fire), but it's concievable that one of the other ghosting
fixes or Eqzitara's fixes got it as well.