Betrayer: Add hotkeys 5,6,7,8,9 & 0 to change UI depth
I default to 0.995 which is pretty far into the scene, but feels more
comfortable to look to the UI from the scene and back again than lower values.
1.0 is too far and feels behind everything (because by definition it is).

I recognise that others may not be used to the UI being so far back, so I've
added a bunch of presets that other people can use.