Book of Unwritten Tales 2: Don't adjust recipe cure page depth in chapter 2
Because the recipe cure (when examined from the inventory) uses the same font
as the menu book (which we already had to disable adjustments for), we can't
adjust the page background for the recipe otherwise the text and page will be
at different depths.

Exclude the texture for the recipe page from being adjusted.