The Book of Unwritten Tales 2: Add convergence preset keybindings
Because F6 causes the game to bug out, so manually adjusting the convergence
with the default keybindings is not a good idea.

L sets a reasonable convergence for most of the game
K sets a lower convergence for certain scenes that display something very close
to the camera, such as the drunken punch minigame at the start of chapter 2.