World of Diving: Use Aion profile
The Aion profile fixes the water and many (but not all) of the haloing issues
in the game, making some of my fixes redundant.

It does break the working shadows, but I've become convinced that they were
working by a fluke anyway and would have broken as I manually fixed some of the
halo issues.

Remove the redundant fixes and set the Aion profile in the ini (still best to
use nVidia inspector)