nVidia Inspector: Name 3D Notes and convergence attributes
nVidia Inspector seems to have some bugs reading string fields, so I'm not 100%
sure that the 3D Notes fields are correct - some games seem to have 3D notes in
those fields, others have gibberish. Editing it doesn't seem to work, and it
wouldn't surprise me if they are all wrong.

Convergence was confirmed by comparing the value saved in the Miasmata profile
with the value we force in DX9Settings.ini, then setting it to different values
(such as 0) and confirming that it changed in game.

Show 3D Notes was determined by reverting all custom settings and noting that
the notes turned on, then turning them off in-game and noting which attribute
got set to 0x00000000. Further experimentation confirmed this was the correct
attribute.