shader database: Use URLs for filenames - easier to look at
My only concern with this method is files being orphaned or updated.

Orphaned files probably isn't a huge deal (chances are they will appear
in an updated zip which I can tie to the original game) - might dump out
an index of these.

Updates are a concern (where the URL doesn't change) - I might want to
add the last modified timestamp to the filenames, but I'll worry about
that later.