Montague's Mount: Fix lens flare
Turns out the lens flare is not only broken, but also uses the same vertex
shader as part of the UI, so we need to do two different fixes in the one
shader depending on what is being drawn.

If the lens flare texture is matched, undo the stereo correction formula, than
adjust it to a fixed depth. This is an approximation, but it looks pretty good
to me.