Montague's Mount: Fix bloom
This is annoying - the fix I did for the lens flare also broke the
bloom, but they use the same texture and shaders, so there wasn't an
easy way to distinguish them apart.

For now I'm matching on depth - the lens flare is drawn close to 0
(seems to be between 0.05 and 0.1 depending on the andle), while the
bloom is drawn at the depth of the lights it is around.