Dreamfall Chapters: Partial fix for broken light at start of game
This light has a double halo effect, of which this only fixes one halo.

The bump map on the ground is misaligned and clips at the edge of screen
- likely related to the second halo?

Shadows line up with the ground, but are not cast from the right origin.

Press U to toggle uncorrecting the vertex shader and O to toggle Mike's
position calculation in the pixel shader (either has the same effect on
one of the halos, but only the vertex shader alters the shadows).