Dreamfall Chapters: Add Shader headers from Unity
These were found by extracting the shaders from the game's asset files
and comparing the assembly. The source code for this can be downloaded
from the Unity website and compiled in Unity, so alternatively a
compiled version of that could be used to identify the shaders (note
that since this particular shader is marked "Hidden", it does not show
up in the Unity UI, making it less straight forward to get Unity to
compile it).