Dreamfall Chapters: Adjust light vertex shader to fix light clipping
This also adds comments to each instruction  to explain what it is doing
from the source code downloaded from the Unity website.

The trick here was that the ray output needed to be adjusted as well as
the UV screen position output. This uses the same maths as I already did
in the pixel shaders, but with a clearer explanation now that I am more
familiar with the maths:

- We are adjusting a view-space coordinate, not a projection-space
  coordinate.

- So, what we are really doing is adjusting the stereo correction
  amount in projection-space by the inverse of the projection matrix to
  get the amount we need to adjust in view-space coordinates.

- The reason we need the horizontal FOV is just a consequence of that
  being used in the fields of the projection matrix we need to do the
  maths (and the other fields involved are all 0, which simplifies most
  of the remaining maths out).

- We divide by P[0,0] because we really want to multiply by P.I[0,0] and
  the maths just happens to work out that P.I[0,0] = 1 / P[0,0]

It seems I still need the correction in the pixel shaders as well to
make the shadows line up. I wouldn't be surprised if I can simplify that
in the vertex shader, but I don't fully understand why I need it twice
yet and wanted to commit this before I dig deeper (perhaps I just missed
a multiplication somewhere, or maybe there is another shader involved
that needs to be corrected, perhaps Internal-PrePassCollectShadows?).