Dreamfall Chapters: Comment SPOT SHADOWS_OFF shader variant from source code
This is just to help me understand what is what in the shader to try to
work out if there's anything else here I need to fix, but maybe it will
help someone else. The source code of the shaders can be downloaded from
the Unity website.

Note that I chose a SHADOWS_OFF variant for the moment since the
assembly is much simpler than a SHADOWS_NATIVE variant (and I have
shadows working).