export_unity_shaders: Use a bitmap to show which headers come from which variants
An extreme case as an example:

// Matched 180 variants of 1 shaders: Hidden/Internal-PrePassLighting

Here we see this particular shader had 180 variants, which all produced
the same assembly (otherwise they wouldn't be grouped in the first
place). Usually you won't get 180 variants, so normally you would expect
much shorter bitmaps and less indentation ;-)

...

// 000000000000000000000000000000fffffffffffffff:     Pass 1/3 {
// 000000000000000fffffffffffffff000000000000000:     Pass 2/3 {
// fffffffffffffff000000000000000000000000000000:     Pass 3/3 {

Now we can clearly see that the shader has three passes and the bitmap
makes it obvious that 1/3 of the variants belong to each pass.

...

// 000000000000000000000000000000fffffffffffffff:       Blend DstColor Zero
// ffffffffffffffffffffffffffffff000000000000000:       Blend One One

We can also see that different variants used different blend modes.
Matching the bitmap up with the above it is clear that all the Pass 1
variants used "Blend DstColor Zero" while Pass 2 and 3 used "Blend One
One".

...

// 000000000000002000000000000002000000000000002:           Keywords { "DIRECTIONAL" "SHADOWS_OFF" "SUNSHINE_DISABLED" }
// 000400000000000000400000000000000400000000000:           Keywords { "DIRECTIONAL" "SHADOWS_OFF" "SUNSHINE_FILTER_HARD" }

And here we start seeing which variants used what keywords - indeed, we
can see that each possible combination of keywords was used in all three
passes.

... and so on