extract_unity_shaders: condense keywords to shorten output filename
Windows explorer is so retarded that it can't deal with a long filename
that is not only perfectly legal in NTFS, but that Windows allowed to be
created in the first place and deals with perfectly fine from the
console!

Reminds me of the old days when we used to protect files on Windows by
using the DOS prompt to add a character in the filename that Explorer
couldn't handle. Now we just need to use really long filenames instead.