Add C++ project to assemble shaders
The idea is to assemble shaders just as a Unity game does so that we can
calculate it's CRC32 and match it up with the shaders dumped from Helix
mod.

Just drag + drop a shader to the exe and it will write out the assembled
shader blob. It doesn't compute the CRC32 itself - it's not difficult,
but it's even easier in one line of Python ;-)

Actually, I'm kind of interested in giving ctypes a shot and seeing if I
can call this from pure python instead.