extract_unity_shaders: Indent shaders like those extracted with Helix mod
This makes the diff between the two cleaner.

There are still some other differences - constants have differing
amounts of significant digits and many of the declarations have a 0
postfix from unity that is not in those from helix mod. None of these
make a difference, so I'm not overly worried.