Dreamfall Chapters: Remove light shaft fix that wasn't working correctly
An example of this effect is at the end of the tunnel in the dark dream.
Talk to the dreamer, turn on the first street lamp then run to the end
of the tunnel.

Without adjusting the VS the effect gets messed up close to the camera,
but the adjustment messes it up far away instead. It does partially fix
the halo, so it's probably along the right track, however I still get a
halo in one eye, so I'm removing the partial fix until I can come up
with something better.

Right now I've been staring at this for too long and need to move on and
come back to this later. I've added all the shaders that I've identified
related to this effect - different pixel shaders are used from different
camera angles, and some only apply when the effect is being drawn
full-screen.

Besides - this effect is glitchy in 2D, so for all I know Red Thread
will change it anyway.