Dreamfall Chapters: LightShift fix WIP
This steroises the light shafts, but I haven't managed to work out how
to fix the one-eyed halo, and this distorts it :(

The _LightPos input is the mono screen-space coords of the light, so
this fix stereoises it in each of the shaders.

U toggles the steroisation in the final interpolation shaders, and P
toggles it in the coords shaders. U makes the rays 3D, and P fixes the
halo in the eye that U just broke, but unfortunately doesn't fix the
halo in the other eye.

At the moment I'm thinking that the output of the coords shader is mono
and needs to be stereoised (I can conditionally disable the whole shader
in both eyes with an if_gt separation, 0 - that should only disable it
in one eye). I haven't had any luck trying to force it to be stereoised
in Helix mod unfortunately.

P also seems to make the light drift sideways in one eye (the one with
no halo) from a distance...

Committing this now... if I can't find a good fix I'll add a toggle to
disable the whole effect before release.