Dreamfall Chapters: Fix water surface (unusual - found slight variation in asset files)
This fix undoes the correction to the reflection coord output. Doing so
makes the reflection look flat on the surface of the water, so there may
be a better fix elsewhere.

This was another shader that didn't match anything in the asset files,
however on closer examination I discovered that it is almost an exact
match for one of the Dreamfall Water FX shaders, but this one has a fog
output that is not in that... my guess is that Unity can add fog outputs
sometimes, so I may have to adjust my scripts to account for that.