Dreamfall Chapters: Fix cascaded directional shadows
The game was using a replacement for
Hidden/Internal-PrePassCollectShadows, which uses a fudged world
position to precompute the camera->world->shadow translation, relying on
the fact that the rays from a directional light are parallel to simplify
the maths.

Unfortunately, using this fudged world position just doesn't work with
the stereo correction formula applied, because we need the correct depth
to remove the stereo correction to get the world position, and we only
get the correct depth in the pixel shader when we sample the camera
depth texture.

This patch replaces the replaced shader with the original from Unity,
and remaps the inputs to match what the original expects. It uses Helix
mod to copy some inputs from the VS to the PS, and others are
transferred as extra outputs of the VS.

The SHADOWS_SCREEN variant of Hidden/Internal-PrePassLighting shader
involved is altered to ignore the stereo correction in the VS and use
screen coordinates directly.

The FOV is still fetched from a different shader that has matching MVP
and MV matrices.

TODO:
- Go through remaining PrePassCollectShadows pixel shaders and replace
  and remap them to the new inputs.
- Examine the extra input that the pixel shader was using that isn't in
  the original and see if we can re-implement it.