Dreamfall Chapters: Pre-emptive fix for another variant of PrePassCollectShadows
NOTE: Other two variants are not fixed! There simply isn't enough usable
ways to transfer the inputs from the VS to the PS, so they can't be
fixed with this technique alone, and will be horribly broken due to the
substituted vertex shader.

I'm leaving them broken as opposed to disabled so that they are obvious
if they are used in-game. The two shaders are 691EDF7D and 8FE84219.

If they become a problem later on I'll have to consider alternatives,
like adding DX9 support to 3DMigoto with unlimited constant
transference, or just modding the unity .asset files directly.