Use implicit constant passing to potentially fix other PrePassCollectShadows variants (UNTESTED!)
bo3b mentioned that Mike figured out that constants defined in a vertex
shader can also be used in the pixel shader, provided that they are not
redefined to something else.

Using this I'm able to implicitly pass constants, which gives me enough
room to pass everthing I need in to pre-emptively fix these variants.

This has not been tested.