Dreamfall Chapters: Fix doubled reflection on floor in Zelenka's Office
The problem stems from the reflection being rendered in mono, and the
real fix would be to force that to render in stereo, however I have not
had any luck doing so.

Instead, fix the glass pixel shader to use vPos and the render target
size instead, which will be correct in both mono and stereo.

I had to pass the _ScreenParams from the vertex shader to the pixel
shader to get this to work, since the DefPSViewSizeConst passed from
Helix mod will only be correct when rendering to the screen.