Dreamfall Chapters: Pre-emptively remove vPos from other lighting shaders
Hopefully will kill any other 1px halos before we encounter them.

I'm wondering if there is a reliable means to test if we are drawing a
full-screen quad in the vertex shader to avoid doing the stereo
correction in the first place... I'm thinking there probably is, because
it looks like it already does this test.