Dreamfall Chapters: Tweak light shafts on menu
My partial fix was less accurate on these than the variant I dialled in
manually. I noticed a relationship between the convergence and the
approximate correct fudge value, so these are now closer for any
convergence. Maybe a similar tweak will work on other light shafts, but
I really need to move on.

I'm thinking the best approach might be to fix this in Cg and mod the
fixed version into Dreamfall - at least that way we could just pass the
depth of the light in instead of having to do all these elaborate
calculations!