extract_unity_headers: Do a better job of combining headers in some circumstances
Before we were prioritising dumping out headers that occurred on the
current line being processed and not the proceeding one, but we did not
consider any lines after that.

Sometimes this could cause a line from a header to be dumped out even
though the same line occurred later in another header we were combining.

The result was occasionally we got a mess as the output - technically
correct, but hard to decipher.

This patch changes the logic to delay outputting a header line if the
same line occurs in any future line. The result is lines like "Program
vp" that are the same in every variant will act as synchronisations
between the different headers so that all previous output will have been
flushed.

This no longer prioritises grouping the same keywords together - if that
proves to be an issue I'll fix that separately.