shadertool: Add automatic fix for common vertex shader halo type issue
This is an attempt to automatically fix vertex shaders that are broken
in a very common way, where the output position has been copied to a
texcoord.  This is not a magic bullet - it can only fix fairly simple
cases of this type of broken shader, but may be useful for fixing halos,
Unity surface shaders, etc.