shadertool: --add-unity-autofog argument to match Unity fog instructions
Some shaders dumped from Unity games with Helix mod have additional fog
related instructions that are not found in the assembly in the .asset
files. It looks like these instructions are being automatically inserted
by Unity (or DX? nVidia? Not really sure) at run-time, but they follow a
pretty simple pattern that is easy to replicate.

The idea is extract_unity_shaders will call this for every extracted
shader to generate fog variants with the correct CRC.

Only vs_3_0 shaders are supported for now.