World of Diving: Use lighting shader for FOV after all
There were a couple of places in the tutorial level where none of the
other shaders I'm grabbing the FOV from were active, so shadows could
become misaligned if the scuba-scooter or camera was used and the FOV
changed.

Re-enable grabbing the FOV from the lighting shader, which seems to work
in those situations.

I'm concerned this shader sometimes may yield a bogus FOV when only
directional lighting is active (previously seen this in the Brony level
before diving), but right now it seems to work.

I've also spotted that the projection matrix is directly available in
the plankton shader in this game, but that is only available underwater.