shadertool: Fix shader model 3 conversion for pixel shaders with colour inputs
This wasn't handling pixel shaders with colour inputs in the form of
"dcl v0.xy", and could produce a shader with the same input defined
twice if it took both a colour and a texcoord input.

Texcoord inputs will now search for a free v* variable rather than using
their texcoord numbers to avoid clashes with existing colour variables.