Add screenshot_archive.py to manage 3D screenshots
I wrote this ages ago after finding out the hard way that the driver
stops taking 3D screenshots after taking 100 of the same game (despite
the fact that it keeps printing out a "screenshot saved" message).

It monitors the screenshot directory and periodically moves new
screenshots into a directory for the game they are from and renames them
to use the date and time they were taken so they will always be in
chronological order.