BoUT2: Add preset to improve recipe book & unadjust background
Remove the stereo correction from the background - it is way too far
back.

Set a predefined separation (15%) and convergence to bring the
foreground close to the screen and not look too disconnected from the
flat background.

Pop the UI 20% * separation out of the screen so it is in front of the book.

Figured out how to use TexCounterReg and [TEX] index to do an adjustment
for only specific textures in the list.