World of Diving: Use projection matrix from Plankton shader whenever active
The Plankton shaders have the projection matrix, but they are only used
while underwater.

Set these shaders to activate preset E, which sets Const2 to 1 and gives
me something I can test in the lighting shaders to determine if the
projection matrix is currently valid. If it is valid I use it, otherwise
I use the MV+MVP matrices as before.

This fixes a couple of spots (mostly in the Diving Base) where no shader
was available with both the MV+MVP matrices, which could muck up the FOV
when getting the camera out or putting away the Scuba-Scooter.

I also tried ResetConst2AfterSet instead of the profile switch, but my
concerns over it's reliability turned out to be justified as Const2 was
always 0.