World of Diving: Whitelist certain UI elements that should always be at depth
Blacklisted textures are never adjusted - used for the diving computer.
Unfortunately the texture CRC for the text changes so it cannot be
blacklisted, and we disable the UI adjustment while tab is held instead.

This mucks up other things in the scene which we want at depth.

So, whitelist those items, using a [TEX] section to set their
TexCounterReg value to 2. Then, we draw them at depth, even when the tab
key is held.