Legend of Aethereus: Guess if drawing a full-screen quad in lighting vertex shader
Some of the pixel shaders are used from a distance with the light drawn
in world, and at other times up close as a full-screen quad. The
adjustment in the vertex shader needs to be applied only when drawn in
world. In other Unity games different lighting shaders were used, but
here the same are used for both.

Detect if drawing a full screen quad in the vertex shader and remove all
undoes from the pixel shaders.

I occasionally see a small amount of minor flickering with this - it's
pretty minor, but there might be a more accurate way to achieve this.