Legends of Aethereus: Autofix all vertex shaders found in game files
This is like doing an autofix over the Dumps/AllShaders/VertexShader/*,
but on steroids - it's run over every vertex shader found in the game
asset & resource files.

This should allow shaders to be fixed that are in levels I haven't even
reached yet.

Same caveat applies as to any autofix - this may break some effects, but
will probably save loads of time overall, so it's probably worth it.