Legends of Aethereus: Stereoise character portraits (inventory, saves, apartment)
Thanks to the Unity surface fixes, rendering this in 2D ends up with a
bad halo (wouldn't it be great if Helix mod would let us turn off the
fixes while rendering to a mono render target? But I digress).

Force the render targets for the character portrait to stereo. This
gives a 3D character in the inventory, and the portraits in the
apartment & load game screen are still mono, but they are from one eye
so no longer have a halo.