Unity: Pass calculated FOV from vertex shader to pixel shaders
While this saves some cycles in the pixel shaders, that is not the
primary motivation for this.

Experience has shown that I cannot rely on getting the FOV from any one
source in Unity games. Sometimes I can grab it from the PrePassLighting
vertex shader, sometimes I can't. Sometimes there's a shader that
already has the projection or inverse projection matrix, though this is
extremely rare. Sometimes I might have no other option than to fudge it.

In many cases I can simply alter the ini file, but sometimes I've found
that is not flexible enough and I may need to choose from several
sources on the fly, in which case I'd have to modify every pixel shader
to select the right source.

This change means I won't have to alter as many shaders to make this
kind of tweak in future - I will only have to change the vertex shaders.

This change is only to the PrePassLighting shaders for now.