Unity: Use matrices from light vertex shader whenever valid
If we are drawing a light in the scene we can be sure that the matrices
passed into the lighting vertex shader will be valid, so we may as well
use them directly.

If we are drawing a full-screen quad we can be sure that they will NOT
be valid, and we *have* to find an alternate source.