DreadOut: Update lighting fix
It seems that we have to use the matrices from the lighting shader most
of the time - they seem to be the only reliable way to fix shadows
outdoors and also work if the phone camera is active.

After entering the school some lights are only visible with the phone
camera. I can't find a good way to fix these, so I've hardcoded the FOV
in the lighting shader whenever a full-screen quad is being drawn, which
seems to have done the trick.