Eleusis: Fix drop shadow on adjusted UI hand icon (NEW UI ADJUSTMENT FORMULA)
This was an odd one - the hand icon in game was adjusted by two shaders,
one made it stereo, the other moved it to the right. However, as it was
pushed further back it's drop shadow became misaligned and moved too far
to the right.

Blacklist it from the shader that moved the cursor to the right, and
change the formula used in the one that adjusts it's depth to alter the
entire coordinate instead of just it's X component.

Normally the UI adjustment formula would be this (The divide by 1 is the
perspective divide and assumes W=1):

S * UI / 1

The normal stereo correction formula is this (including the perspective
divide, which we normally wouldn't apply manually):

S * (W - C) / W

So, in order to find an adjustment we can make to the output position
that would use the stereo correction formula to find the equivalent
depth to the UI adjustment formula, we set them equal:

S * (W - C) / W = S * UI / 1

And solve for W:

W = -C / (UI - 1)

Then all we do is multiply the output coordinate (all four components)
by that value.