Eleusis: Kill Const3 debugging keybinding
Hitting this by accident will kill the UI since it pushes it to
infinity, which doesn't work with the new UI depth adjustment formula.
I only use this keybinding for debugging, so just kill it.

Remap Const2 debugging keybinding to O, since I is used for Inventory in
this game.