Stranded Deep: Partial water fix (pixel shader variant)
These issues can be fixed in the vertex shader, however in this game it
is already using all four sampler registers so I am unable to use it.

Instead do the fix in the pixel shader - this fixes refraction and halo
issues. A halo is still visible on the light pattern, but I think there
is another shader responsible for that.