Stranded Deep: Partial fix for shadows and caustic highlights underwater
These effects are applied in the water fog volume shader, which works
completely differently to standard Unity lighting.

It uses the world coordinates of the view frustrum and camera position,
so I convert them into view-space coordinates where I can apply the
view-space stereo correction formula.

Helix mod is used to copy the MV.I and MVP matrices in to calculate
P.I[0,0] for the view-space correction like we do for Unity shadows.

We also take the inverted view matrix in this shader and invert it back
to a regular view matrix and pass it back in.

Shadows & caustic highlights are much closer with this, but are still
not perfectly lined up with the ocean floor, so I must still be missing
something, but since they are distorted by the water surface it's close
enough for now.