Stranded Deep: UseEndScene = false to stop debug text messing up water
Normally I don't care if the debug text ends up projected on something
in the scene since it can easily be turned off by key press, but in this
game it not only messes up the rendering, it also messes up the wave
physics. So much so that loading a game with it on will have completely
bonkers water *everywhere*, including on land.

UseEndScene=false resolves the issue, and the text still works.