Stranded Deep: Change Crepuscular ray fix to not break sun shafts
The same vertex shader is used for both underwater Crepuscular rays and
sun shafts, but fixing it in the vertex shader only works for the
underwater rays, and not the sun shafts on land.

Instead, add an extra output from that vertex shader with the corrected
output leaving the default output unchanged. Change the pixel shader to
use the corrected output.