Miasmata: Correct fix for high quality water
I've determined the issue with the real time reflections was due to an
uncorrected world-space coordinate. Pass a corrected version as an extra
output from the vertex shader to use it where needed in the pixel
shader.

I wasn't able to unproject the final coordinate, as this proved unstable
- so I just unproject the correction amount, which is perfect. Some
lines in the pixel shader still need the uncorrected coordinate, so
interleave the instructions using corrected and uncorrected coordinates
where needed.

There still looks to be an issue far on the horizon where the
reflections are not at the right coordinates.