Miasmata: Fix light shafts (new technique)
This is the vertex shader for the final stage of the light shafts in
this game, where they are transferred from a radial coordinate system
into screen coordinates.

The vertex shader has an input which I believe is the world coordinates
of the vertex being operated on, which need to be corrected. Problem is
- we have no way to get the correct amount to adjust it by in world
coordinates. As such, we take the world-view-projection matrix of an
object that is never rotated, in this case we just use the water matrix
since we already have it. We run two points back through this shader and
subtract the resulting points, which removes the objects translation
leaving the correction amount in world space.