Miasmata: Improve blending of water shader when near edge of map
Seems this shader is responsible for drawing the water right at the edge
of the map which appears to be on the far horizon. In two parts of the
map it is possible to see more detail on the edge of the map and this
didn't appear to blend well with the rest of the water. Using the
uncorrected coordinate in this part makes it look better - it's still
not great as the fact that it is basically a wall is apparent in 3D, but
it blends a lot better.