Miasmata: Hide clipping on edge of light shafts
Similar to how objects disappear beyond the mono frustrum (ie at the
left & right beyond the edge of the screen) in 3D, so does the adjusted
light shafts and fog. In this case the effect can be quite jarring, as
the fog will be totally transparent in section of a sphere.

To improve matters, pass both corrected and uncorrected world positions
to the pixel shader, and use vPos to determine when it is outside the
bounds of the screen by more than 1/2 separation and use the uncorrected
coordinate instead.

This can make the light shafts repeat at the edge of the screen, but it
is barely noticeable, and much better than large sections of the sky
being a mismatched colour.

Something a bit strange is going on with passing the extra output to the
pixel shader - trying to pass the corrected coordinates in the new
output (with no other change) resulted in clipping on *all four* sides
of the screen. I can only assume there are some differences in the
interpolation settings for the outputs. As a result I pass the corrected
coordinate through the original output, and the uncorrected coordinate
through the new one.