Miasmata: Hide the workaround that hides the clipping fog/light shafts
Interpolate between the corrected and uncorrected coordinates in part of
the screen to hide the fact that we are hiding the clipping.

This divides the screen into five sections - a full stereo section in
the middle, two mono sections on the outside and two interpolation
sections between them.

     --------------------- 1/2 separation - clipping occurs here
    /
   /                  ---- separation width
  /                  /
<->                <--->
------------------------
| | |              | | |
| | |    SCREEN    | | |
| | |              | | |
------------------------
 ^ ^      ^^^^^     ^ ^
 | |        |       | |
 | |        \-------+-+--- Uses corrected coordinates for stereo light shafts
 | |                | |
 | |                | \
 \-+----------------+----- Uses uncorrected coordinates to hide clipping
   |                |
   |                \
   \---------------------- Interpolates between uncorrected and corrected coordinates

Also spotted and fixed one place I hadn't switched to use the corrected
coordinates, which might explain the weird distinction I was seeing
between which output I was using earlier.