Miasmata: Reduce interpolation on light shafts to outer separation/2 of screen
Experimenting further with the interpolation approach to hiding the
clipping on the fog/light shafts proved that it is only necessary to do
this on the outer 1/2 separation of the screen. This still removes the
jarring clipping, but means that the light shafts only bend right out at
the side of the screen, which is less noticeable than over the whole
separation at the edge, so this is better.