shadertool: Add --disable-redundant-unreal-correction auto fix mode
This option will look for pixel shaders that have a vPos semantic and
use the Unreal Engine NvStereoEnabled constant, and disable the later.

This has the same effect as disabling stereo 3D support through
BaseEngine.ini for these shaders, but such that it remains possible to
continue using that for other reasons (CinderellaMan reported 3D would
not engage at all with that option set to False).