Miasmata: Add key to lower separation to work around falsely obscured landmarks
It seems landmarks are matched on the screen by their mono position, but
I think they are using the left eye of a stereo depth buffer to test if
they are obscured. As such, a landmark can be falsely obscured by
another object just to it's left. I haven't had any luck playing with
render targets yet, so the next best workaround is a hot key to
temporarily minimise separation (disabling 3D also works).