Miasmata: Add automatic stereo cursor
I've identified a sampler used by the water shader that contains (among
other things) the scene depth. Copy this into the UI shader and use it
to automatically adjust the cursor depth based on the depth buffer.

At the moment this only samples the depth buffer in the center of the
screen, which is not ideal - but close enough to start. It would
probably be better to loop over a number of points searching for an
object hit, which may improve the stereo depth near the edge of an
object.

One important gotcha to note (took me a while to realise this) - as the
UI shader has W=1, it will not have a perspective divide applied, but
the rest of the scene does. To make the stereo cursor match up with the
scene we must therefore apply the perspective divide ourselves.